refactor: sliderのコードをリファクタリング #16
